[NEW]: LaTeX, lists and references
The lexor converter underwent major changes. These changes allowed a
more simplified way of writing the converting styles. There are a few
more things to cover with this node-converters. For instance, when
including a file that contains references from the outside file, then
the reference will not get updated. When including files we need to
check for other options and verify that the file is available (may want
to put an option to write a warning in the file). The python converter
also needs to be rewritten.