[FIX]: include and python embeddings
The include node parser behaves more like the include function for
python embeddings which have now been activated. The include node has
certain attributes that can be modified. Information on this will have
to be written in the documentation. For now look over the definition of
IncludeNC to see the attributes. The converter function has a new
function that has been added to it. This function is essential to the
include functionality. Unfortunately this function is buggy right now
and all the attributes such as the names of languages must be spelled
out in full.