[NEW]: documentclass and usepackage
Taken from LaTeX, lexor can now use the tag `usepackage` to specify
that a python script should be run for the document. This python script
is a module which defines the function `convert` with the arguments
`converter` and `document`. The `documentclass` tag tells us that we
will be using an html document and thus it inserts the html and head
tag into the document in its place. A few more changes were made into
the other modules to accommodate the function `remove_node`.