[NEW]: meta
We can now include meta information in our documents. By doing so we
can access the meta dictionary of the `Document` object. A few
rearrangements were done to include and document. `DocumentClassNC` now
checks to see if we have a valid HTML document. It shouldn't affect if
we already declared the html, head and body elements. The
`UsePackageNC` now uses the meta information to run additional packages.