[NEW]: codeblock, latex patch
The code block converter has been added. This is still work in progress
as there are many syntax highlighter packages out there. The goal of
this converter will be to adapt the code to whichever package the user
desires. For now the parser and converter have been adapted to work
with <http://alexgorbatchev.com/SyntaxHighlighter/>. The element tag
latex has not been declared as a `RawText` element. This is will done
in another release. For now the changes to `latex.py` fixes issues in
some documents.