[BREAKING]: guinea pig
The converter module in lexor is undergoing breaking changes. Since
this is the only converter things have broken here beyond repair. This
is the point where we build everything again. For now we can take a
look at the inline module which is pretty much set. But then again,
things may change, that’s why we are writing the tests.